+++ driver-core/src/test/java/com/datastax/driver/core/SchemaTest.java
@@ -113,7 +113,7 @@ public class SchemaTest extends CCMBridge.PerClassSingleNodeCluster {
     // way to check we correctly handle schemas so it's probably not so bad.
     // In particular, exportAsString *does not* guarantee that you'll get
     // exactly the same string than the one used to create the table.
-    @Test(groups = "long")
+    @Test(groups = "short")
     public void schemaExportTest() {
 
         KeyspaceMetadata metadata = cluster.getMetadata().getKeyspace(TestUtils.SIMPLE_KEYSPACE);
@@ -132,7 +132,7 @@ public class SchemaTest extends CCMBridge.PerClassSingleNodeCluster {
     }
 
     // Same remark as the preceding test
-    @Test(groups = "long")
+    @Test(groups = "short")
     public void schemaExportOptionsTest() {
         TableMetadata metadata = cluster.getMetadata().getKeyspace(TestUtils.SIMPLE_KEYSPACE).getTable("with_options");
         assertEquals(metadata.exportAsString(), withOptions);
